% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ingredients.R
\name{add_ingredients}
\alias{add_ingredients}
\title{Add ingredients}
\usage{
add_ingredients(cmap_data, small_molecules, ctd_data, np_data, lincs_data,
  new_data = NULL)
}
\arguments{
\item{cmap_data}{Data from CMAP2 from the Harmonizome}

\item{small_molecules}{Data from small molecules treatments on cell lines obtained from the Harmonizome}

\item{ctd_data}{Data from CTD, filtered on mRNA effects from drugs}

\item{np_data}{Data on MCF7 cell lines treated with natural products from traditional chinese medicine}

\item{lincs_data}{Data from LINCS L1000 from the Harmonizome}
}
\value{
A list of TF-IDF matrices and respective cross-product vectors.
}
\description{
This function takes in data from multiple sources and generates a drug-term matrix as required by `DRUID`.
}
