% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alchemy.R
\name{alchemy}
\alias{alchemy}
\title{Alchemy}
\usage{
alchemy(add_new = NULL)
}
\value{
A list of TF-IDF matrices and respective cross-product vectors.
}
\description{
This function is a wrapper to the `cauldron` package. It will generate the TF-IDF matrices and cross-products that are used in DRUID. As of this version, the functions in `cauldron` are self-contained and take no input. This will change in future releases.
}
