% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod_matrix.R
\name{crossprod_matrix}
\alias{crossprod_matrix}
\title{Cross-product vector for the TF-IDF}
\usage{
crossprod_matrix(tfidf_matrix)
}
\arguments{
\item{tfidf_matrix}{Sparse TF-IDF matrix}
}
\value{
cross-product of the TF-IDF matrix
}
\description{
Returns the cross-product for an input TF-IDF matrix
}
