% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid_potion.R
\docType{data}
\name{druid_potion}
\alias{druid_potion}
\title{DRUID potion}
\format{An object of class \code{list} of length 5.}
\usage{
druid_potion
}
\description{
A set of tf-idf matrices from multiple sources to be used in DRUID.
}
\keyword{datasets}
