% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu}
\alias{menu}
\title{Drug menu
Builds the drug metadata data frame, from the data sets used to build the compendium.}
\usage{
menu(ingredients)
}
\arguments{
\item{ingredients}{Data frames and metadata for the drugs in each of the data sets to be compbined.}
}
\value{
A named list with all the drugs sets
}
\description{
Drug menu
Builds the drug metadata data frame, from the data sets used to build the compendium.
}
