% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_pot.R
\name{mix_pot}
\alias{mix_pot}
\title{Mix the pot
This function will calculate the TF-IDF matrix and its cross-product for each of the drug compendia added. The TF-IDF will be calculated using the \link{\code{ctfidf}} function and the cross-product with the \link{\code{crossprod_matrix}} function.  The TF-IDF and its cross-product will be determined independently for each data set.}
\usage{
mix_pot(bowl_list)
}
\arguments{
\item{bowl_list}{A list of data sets (compendia), from the \link{\code{add_ingredients}} function.}
}
\value{
A TF-IDF data matrix and a crossproduct vector for each of the data sets.
}
\description{
Mix the pot
This function will calculate the TF-IDF matrix and its cross-product for each of the drug compendia added. The TF-IDF will be calculated using the \link{\code{ctfidf}} function and the cross-product with the \link{\code{crossprod_matrix}} function.  The TF-IDF and its cross-product will be determined independently for each data set.
}
