% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_recipe.R
\name{pour_recipe}
\alias{pour_recipe}
\title{Pour recipe}
\usage{
pour_recipe(pot_contents, menu_items)
}
\arguments{
\item{pot_contents}{TF-IDF and cross-product vector for drug matrices, as computed with the \link{\code{mix_pot}} function}

\item{menu_items}{Drugs metadata as obtained from the \link{\code{menu}} function}
}
\value{
A named list with TF-IDF, cross-product vectors, and drug metadata for all of the data sources used.
}
\description{
This functions will put all the TF-IDF matrices together, as well as a combined TF-IDF will all data sources.
}
