% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talaria_io.R
\name{TALARIA_from_seurat}
\alias{TALARIA_from_seurat}
\title{Convert a Seurat object to a SingleCellExperiment}
\usage{
TALARIA_from_seurat(sobj, assay = "RNA", verbose = TRUE)
}
\arguments{
\item{sobj}{A \code{Seurat} object.}

\item{assay}{Character. Which Seurat assay to convert.  Default
\code{"RNA"}.}

\item{verbose}{Logical. Print a summary. Default \code{TRUE}.}
}
\value{
A \code{SingleCellExperiment}.
}
\description{
Wraps \code{Seurat::as.SingleCellExperiment()} to convert a \code{Seurat}
object to a \code{SingleCellExperiment}, preserving assays, embeddings,
and cell metadata.
}
