% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talaria_logging.R
\name{TALARIA_get_log_file}
\alias{TALARIA_get_log_file}
\title{Get the current log file path}
\usage{
TALARIA_get_log_file()
}
\value{
Character path to the active log file, or \code{NULL} if no log
is running.
}
\description{
Get the current log file path
}
