% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talaria_logging.R
\name{TALARIA_is_logging}
\alias{TALARIA_is_logging}
\title{Check whether logging is active}
\usage{
TALARIA_is_logging()
}
\value{
Logical — \code{TRUE} if a log is currently running.
}
\description{
Check whether logging is active
}
