% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAULDRON-package.R
\docType{package}
\name{CAULDRON-package}
\alias{CAULDRON}
\alias{CAULDRON-package}
\title{CAULDRON: Single-Cell and Single-Nuclei RNA-seq Analysis Toolkit}
\description{
Modular pipeline for single-cell and single-nuclei RNA-seq analysis including quality control, preprocessing, dimensionality reduction, clustering, annotation, differential expression, trajectory analysis, cell-cell communication, and visualization. Built on SingleCellExperiment (Bioconductor). Named after figures from Norse mythology.
}
\author{
\strong{Maintainer}: Yohan Lefol \email{email@example.com}

}
\keyword{internal}
