% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_recipe.R
\name{pour_recipe}
\alias{pour_recipe}
\title{Pour recipe}
\usage{
pour_recipe(pot_contents, menu_items)
}
\value{
A named list with TF-IDF, cross-product vectors, and drug metadata for all of the data sources used.
}
\description{
This functions will put all the TF-IDF matrices together, as well as a combined TF-IDF will all data sources.
}
